Updated to check if ID not found.

+++ src/main/java/com/elixr/poc/service/UserService.java
@@ -1,6 +1,8 @@
 package com.elixr.poc.service;
 
+import com.elixr.poc.constants.ApplicationConstants;
 import com.elixr.poc.repository.UserRepository;
+import com.elixr.poc.rest.response.DeleteResponse;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -12,6 +14,7 @@ Service layer
 public class UserService {
 
     private final UserRepository userRepository;
+    private DeleteResponse deleteResponse;
 
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
@@ -20,14 +23,20 @@ public class UserService {
     /*
     Deleting the user by the userId.
      */
-    public boolean deleteUserDetails(UUID userId) {
-        boolean success = false;
-        try {
+    public DeleteResponse deleteUserDetails(UUID userId) {
+        deleteResponse = new DeleteResponse();
+        boolean userRecords = userRepository.existsById(userId);
+        if (userRecords) {
             userRepository.deleteById(userId);
-            success = true;
-        } catch (Exception e) {
-            e.printStackTrace();
+            deleteResponse.setSuccess(true);
+            ApplicationConstants.REQUEST_VALUE = 200;
+            deleteResponse.setMessage(ApplicationConstants.SUCCESSFULLY_DELETED);
+            return deleteResponse;
+        } else {
+            deleteResponse.setSuccess(false);
+            deleteResponse.setMessage(ApplicationConstants.ID_MISMATCH);
+            ApplicationConstants.REQUEST_VALUE = 404;
+            return deleteResponse;
         }
-        return success;
     }
 }
